chat_demo/requirements.txt
git+git://github.com/reflex-dev/reflex.git
reflex-chat
openai
